feat(qwen-proxy): extract upstream types to types.ts with ChatCompletionsBody re-export shim
Move OpenAiChatChunk, OpenAiChatCompletion, Model to src/upstream/types.ts.
Add ChatCompletionsBody interface (shared body shape for chat completions).
Re-export all from client.ts so existing imports resolve unchanged.
UpstreamClient.chatCompletions now uses ChatCompletionsBody param type.
